+++ spec/fixtures/dummy.js
@@ -1,1 +1,1 @@
-console.log('dummy')
+'in dummy'